# api/auth/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/profile": {
            "get": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Возвращает полную информацию о профиле текущего пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Получить профиль пользователя",
                "responses": {
                    "200": {
                        "description": "Профиль пользователя",
                        "schema": {
                            "$ref": "#/definitions/dto.ProfileResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Профиль не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/profile/info": {
            "put": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Обновляет основную информацию профиля пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Обновить основную информацию профиля",
                "parameters": [
                    {
                        "description": "Данные для обновления профиля",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.ProfileUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Профиль успешно обновлен",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Профиль не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/profile/interests": {
            "put": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Обновляет интересы пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Обновить интересы пользователя",
                "parameters": [
                    {
                        "description": "Список интересов",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Interest"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Интересы успешно обновлены",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/profile/photo": {
            "post": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Загрузка фотографий профиля",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Загрузить фотографии",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Фотографии для загрузки",
                        "name": "photos",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная загрузка фото",
                        "schema": {
                            "$ref": "#/definitions/domain.UserPhoto"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Превышен лимит фото",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/profile/photo/{id}": {
            "put": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Устанавливает основную фотографию профиля",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Установить основное фото",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID фотографии",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Основное фото успешно установлено",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Фото не принадлежит пользователю",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Фото не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Удаляет фотографию профиля",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Удалить фотографию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID фотографии",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Фото успешно удалено",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Фото не принадлежит пользователю",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Фото не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/profile/preferences": {
            "put": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Обновляет предпочтения пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Обновить предпочтения пользователя",
                "parameters": [
                    {
                        "description": "Данные предпочтений",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.PreferencesUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Предпочтения успешно обновлены",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "constants.Gender": {
            "type": "string",
            "enum": [
                "male",
                "female"
            ],
            "x-enum-varnames": [
                "GenderMale",
                "GenderFemale"
            ]
        },
        "constants.GenderPreference": {
            "type": "string",
            "enum": [
                "both",
                "male",
                "female"
            ],
            "x-enum-varnames": [
                "GenderPrefBoth",
                "GenderPrefMale",
                "GenderPrefFemale"
            ]
        },
        "constants.InterestType": {
            "type": "string",
            "enum": [
                "workout",
                "fun",
                "party",
                "chill",
                "love",
                "relax",
                "yoga",
                "friendship",
                "culture",
                "cinema"
            ],
            "x-enum-varnames": [
                "InterestTypeWorkout",
                "InterestTypeFun",
                "InterestTypeParty",
                "InterestTypeChill",
                "InterestTypeLove",
                "InterestTypeRelax",
                "InterestTypeYoga",
                "InterestTypeFriendship",
                "InterestTypeCulture",
                "InterestTypeCinema"
            ]
        },
...
        "domain.Interest": {
            "type": "object",
            "properties": {
                "theme": {
                    "$ref": "#/definitions/constants.InterestType"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "domain.PreferencesUpdateRequest": {
            "type": "object",
            "properties": {
                "age_max": {
                    "type": "integer"
                },
                "age_min": {
                    "type": "integer"
                },
                "global_search": {
                    "type": "boolean"
                },
                "max_distance": {
                    "type": "integer"
                },
                "show_gender": {
                    "$ref": "#/definitions/constants.GenderPreference"
                }
            }
        },
        "domain.ProfileUpdateRequest": {
            "type": "object",
            "properties": {
                "artist": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/constants.Gender"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "quote": {
                    "type": "string"
                }
            }
        },
...
        "domain.UserPhoto": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "display_order": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "is_approved": {
                    "type": "boolean"
                },
                "photo_url": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.ProfileResponse": {
            "type": "object",
            "properties": {
                "interests": {},
                "photos": {},
                "preferences": {},
                "user": {}
            }
        },
...
        "dto.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                }
            }
        },
...
    }
}